Add ways to use the notes created for symbols.
1. When the mouse hovers on a symbol that has a note, the note will
be shown in a tooltip.

2. When a new Navigation Frame is created, if the symbol it represents
has a note, it will be attached to the frame. "Edit note" function in
this frame will edit the old note, instead of creating new one.

Other changes to support these features:

1. Changed the backend API, so that (POST, /notes/) will create a new
notes, AND return the new list of notes.

2. Changed the correspond front end function, to update the notelist
after the (POST, /notes/) call.

3. Updated related test cases.